refactor: rename convertResponseToObject to convert
