+++ src/libs/notion/index.ts
@@ -1,38 +1,38 @@
 // export * from "./auth.js";
-export * from "./blocks/convertResponseToBlock/audio.js";
+export * from "./blocks/convert/audio.js";
 export * from "./blocks/retrieve.js";
-export * from "./blocks/convertResponseToBlock/bookmark.js";
-export * from "./blocks/convertResponseToBlock/breadcrumb.js";
-export * from "./blocks/convertResponseToBlock/bulletedListItem.js";
-export * from "./blocks/convertResponseToBlock/callout.js";
-export * from "./blocks/convertResponseToBlock/childDatabase.js";
-export * from "./blocks/convertResponseToBlock/childPage.js";
-export * from "./blocks/convertResponseToBlock/code.js";
-export * from "./blocks/convertResponseToBlock/column.js";
-export * from "./blocks/convertResponseToBlock/columnList.js";
-export * from "./blocks/convertResponseToBlock/divider.js";
-export * from "./blocks/convertResponseToBlock/embed.js";
-export * from "./blocks/convertResponseToBlock/equation.js";
-export * from "./blocks/convertResponseToBlock/file.js";
-export * from "./blocks/convertResponseToBlock/heading1.js";
-export * from "./blocks/convertResponseToBlock/heading2.js";
-export * from "./blocks/convertResponseToBlock/heading3.js";
-export * from "./blocks/convertResponseToBlock/image.js";
-export * from "./blocks/convertResponseToBlock/linkPreview.js";
-export * from "./blocks/convertResponseToBlock/linkToPage.js";
-export * from "./blocks/convertResponseToBlock/numberedListItem.js";
-export * from "./blocks/convertResponseToBlock/paragraph.js";
-export * from "./blocks/convertResponseToBlock/pdf.js";
-export * from "./blocks/convertResponseToBlock/quote.js";
-export * from "./blocks/convertResponseToBlock/syncedBlock.js";
-export * from "./blocks/convertResponseToBlock/tableRow.js";
-export * from "./blocks/convertResponseToBlock/table.js";
-export * from "./blocks/convertResponseToBlock/tableOfContents.js";
-export * from "./blocks/convertResponseToBlock/template.js";
-export * from "./blocks/convertResponseToBlock/toDo.js";
-export * from "./blocks/convertResponseToBlock/toggle.js";
-export * from "./blocks/convertResponseToBlock/unsupported.js";
-export * from "./blocks/convertResponseToBlock/video.js";
+export * from "./blocks/convert/bookmark.js";
+export * from "./blocks/convert/breadcrumb.js";
+export * from "./blocks/convert/bulletedListItem.js";
+export * from "./blocks/convert/callout.js";
+export * from "./blocks/convert/childDatabase.js";
+export * from "./blocks/convert/childPage.js";
+export * from "./blocks/convert/code.js";
+export * from "./blocks/convert/column.js";
+export * from "./blocks/convert/columnList.js";
+export * from "./blocks/convert/divider.js";
+export * from "./blocks/convert/embed.js";
+export * from "./blocks/convert/equation.js";
+export * from "./blocks/convert/file.js";
+export * from "./blocks/convert/heading1.js";
+export * from "./blocks/convert/heading2.js";
+export * from "./blocks/convert/heading3.js";
+export * from "./blocks/convert/image.js";
+export * from "./blocks/convert/linkPreview.js";
+export * from "./blocks/convert/linkToPage.js";
+export * from "./blocks/convert/numberedListItem.js";
+export * from "./blocks/convert/paragraph.js";
+export * from "./blocks/convert/pdf.js";
+export * from "./blocks/convert/quote.js";
+export * from "./blocks/convert/syncedBlock.js";
+export * from "./blocks/convert/tableRow.js";
+export * from "./blocks/convert/table.js";
+export * from "./blocks/convert/tableOfContents.js";
+export * from "./blocks/convert/template.js";
+export * from "./blocks/convert/toDo.js";
+export * from "./blocks/convert/toggle.js";
+export * from "./blocks/convert/unsupported.js";
+export * from "./blocks/convert/video.js";
 export * from "./richText/equation.js";
 export * from "./richText/mention.js";
 export * from "./richText/richText.js";
